Commit #2

+++ dags/churn.py
@@ -21,7 +21,7 @@ def prepare_churn_dataset():
     @task()
     def create_table():
         import sqlalchemy
-        from sqlalchemy import MetaData, Table, Column, Integer, Float, UniqueConstraint, inspect
+        from sqlalchemy import MetaData, Table, Column, Integer, BigInteger, Float, UniqueConstraint, inspect
 
         metadata = MetaData()
         table = Table(
@@ -36,7 +36,7 @@ def prepare_churn_dataset():
             Column('is_apartment', Integer),
             Column('studio', Integer),
             Column('total_area', Float),
-            Column('price', Integer),
+            Column('price', BigInteger),
             Column('build_year', Integer),
             Column('building_type_int', Integer),
             Column('latitude', Float),
@@ -81,7 +81,7 @@ def prepare_churn_dataset():
         # Удаляем строки с пустыми ценами
         data = data[~data['price'].isnull()]
         
-        # Дабавляем target = ln(1 + price)
+        # Дабавляем целевой признак target = ln(1 + price)
         data['target'] = np.log1p(data['price'])
 
         return data
